saran wak

+++ Controllers/Flags_Controller.js
@@ -11,36 +11,41 @@ const upload = multer({ storage: storage });
 
 exports.GetFlags = async (request, response) => {
   /** call findAll() to get all data */
-  let Flags = await FlagsModel.findAll();
-  return response.json({
-    success: true,
-    data: Flags,
-    message: `All Members have been loaded`,
-  });
+  
+    let Flags = await FlagsModel.findAll();
+    return response.json({
+      success: true,
+      data: Flags,
+      message: `All Members have been loaded`,
+    });
+  
 };
 exports.GetFlagsByID = async (request, response) => {
   /** call findAll() to get all data */
-  let id_flag = request.params.id;
-  let Flags = await FlagsModel.findOne({
-    where: {
-      id_product: id_flag,
-    },
-  });
-  return response.json({
-    success: true,
-    data: Flags,
-    message: ` Have Been Loaded`,
-  });
+  
+    let id_flag = request.params.id;
+    let Flags = await FlagsModel.findOne({
+      where: {
+        id_product: id_flag,
+      },
+    });
+    return response.json({
+      success: true,
+      data: Flags,
+      message: ` Have Been Loaded`,
+    });
+  
 };
 exports.InsertFlag = async (req, res) => {
   try {
-    const { Name, Price } = req.body;
-    const imageBuffer = req.file.buffer;
-    if (req.session.isAdmin) {
+    if (req.session.isAdmin || req.user.isAdmin == "True") {
+      const { Name, Price, Continent } = req.body;
+      const imageBuffer = req.file.buffer;
       const newFlag = await FlagsModel.create({
         flags: imageBuffer,
         name: Name,
         price: Price,
+        continent: Continent,
       });
       res.json({
         success: true,
@@ -50,7 +55,7 @@ exports.InsertFlag = async (req, res) => {
       res.json({
         success: false,
         message: "You are not admin",
-      })
+      });
     }
   } catch (error) {
     console.error(error);
@@ -62,24 +67,31 @@ exports.InsertFlag = async (req, res) => {
 };
 exports.DeleteFlag = async (req, res) => {
   try {
-    const ID = req.params.id;
-    const result = await FlagsModel.destroy({ where: { id_product: ID } });
-    await FlagsModel.sequelize.query(
-      "ALTER TABLE `product` AUTO_INCREMENT = 1"
-    );
-    if (result === 0) {
-      // The destroy method returns the number of affected rows, so if it's 0, the record was not found.
-      res.status(404).json({
-        success: false,
-        message: `Flag with ID ${ID} not found`,
-      });
-      return;
-    }
+    if (req.session.isAdmin || req.user.isAdmin == "True") {
+      const ID = req.params.id;
+      const result = await FlagsModel.destroy({ where: { id_product: ID } });
+      const reset = await FlagsModel.sequelize.query(
+        "ALTER TABLE `product` AUTO_INCREMENT = 1"
+      );
+      if (result === 0) {
+        // The destroy method returns the number of affected rows, so if it's 0, the record was not found.
+        res.status(404).json({
+          success: false,
+          message: `Flag with ID ${ID} not found`,
+        });
+        return;
+      }
 
-    res.json({
-      success: true,
-      message: `Flag with ID ${ID} deleted successfully`,
-    });
+      res.json({
+        success: true,
+        message: `Flag with ID ${ID} deleted successfully`,
+      });
+    } else {
+      res.json({
+        success: false,
+        message: "You are not admin",
+      });
+    }
   } catch (error) {
     console.error(error);
     res.status(500).json({
@@ -90,23 +102,31 @@ exports.DeleteFlag = async (req, res) => {
 };
 exports.DeleteAllFlag = async (req, res) => {
   try {
-    const result = await FlagsModel.destroy({ truncate: true });
-    await FlagsModel.sequelize.query(
-      "ALTER TABLE `product` AUTO_INCREMENT = 1"
-    );
-    console.log(result);
-    if (result === 0) {
-      // The destroy method returns the number of affected rows, so if it's 0, the record was not found.
+    if (req.session.isAdmin || req.user.isAdmin == "True") {
+      const result = await FlagsModel.destroy({ truncate: true });
+      await FlagsModel.sequelize.query(
+        "ALTER TABLE `product` AUTO_INCREMENT = 1"
+      );
+      console.log(result);
+      if (result === 0) {
+        // The destroy method returns the number of affected rows, so if it's 0, the record was not found.
+        res.json({
+          success: true,
+          message: `All Flags Deleted Successfully`,
+        });
+        return;
+      } else {
+        res.status(200).json({
+          success: false,
+          message: `No Flag found`,
+        });
+      }
+    } else {
       res.json({
-        success: true,
-        message: `All Flags Deleted Successfully`,
+        success: false,
+        message: "You are not admin",
       });
-      return;
     }
-    // res.status(200).json({
-    //   success: false,
-    //   message: `No Flag found`,
-    // });
   } catch (error) {
     console.error(error);
     res.status(500).json({
@@ -117,32 +137,40 @@ exports.DeleteAllFlag = async (req, res) => {
 };
 exports.UpdateFlag = async (req, res) => {
   try {
-    const { Name, Price } = req.body; // Assuming you also send the ID in the request body
-    const id = req.params.id;
+    if (req.session.isAdmin || req.user.isAdmin == "True") {
+      const { Name, Price, Continent } = req.body; // Assuming you also send the ID in the request body
+      const id = req.params.id;
 
-    const imageBuffer = req.file;
+      const imageBuffer = req.file;
 
-    // Check if the record with the given ID existsxa
-    const existingFlag = await FlagsModel.findByPk(id);
+      // Check if the record with the given ID existsxa
+      const existingFlag = await FlagsModel.findByPk(id);
 
-    if (!existingFlag) {
-      return res.status(404).json({
+      if (!existingFlag) {
+        return res.status(404).json({
+          success: false,
+          message: `Flag with ID ${id} not found`,
+        });
+      }
+
+      // Update the existing record
+      await existingFlag.update({
+        flags: imageBuffer,
+        name: Name,
+        price: Price,
+        continent: Continent,
+      });
+
+      res.json({
+        success: true,
+        message: `Flag with ID ${id} updated successfully`,
+      });
+    } else {
+      res.json({
         success: false,
-        message: `Flag with ID ${id} not found`,
+        message: "You are not admin",
       });
     }
-
-    // Update the existing record
-    await existingFlag.update({
-      flags: imageBuffer,
-      name: Name,
-      price: Price,
-    });
-
-    res.json({
-      success: true,
-      message: `Flag with ID ${id} updated successfully`,
-    });
   } catch (error) {
     console.error(error);
     res.status(500).json({
